Added platform specific file structure
