+++ perfstats/perfstats_unix.go
@@ -1,0 +1,11 @@
+package perfstats
+
+import "fmt"
+
+func SayHi() {
+	fmt.Println("hi unix")
+}
+
+func PlatformSysStats() []byte {
+	return nil
+}
